fixed updating of user, creating of todolists with tasks and unit tests
