+++ src/main/java/com/list/todo/TodoListApplication.java
@@ -26,12 +26,5 @@ public class TodoListApplication extends AsyncConfigurerSupport {
 		executor.initialize();
 		return executor;
 	}
-
-	// TODO: девалидация токена
-	// TODO: придумать как убрать условные конструкции в контроллерах
-	// TODO: добавить профили для конфигурации Spring
-	// TODO: интеграционные тесты и юнит тесты для контроллеров
-	// TODO: выделить общий интерфейс для нотификации, реализацией будет нотификация по емейлу и вебсокетам
-    // TODO: нужна валидация для dto?
 }
