+++ src/main/java/com/list/todo/payload/UpdatingUserInput.java
@@ -1,0 +1,23 @@
+package com.list.todo.payload;
+
+import lombok.Data;
+
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Size;
+
+@Data
+public class UpdatingUserInput {
+    @Size(min = 4, max = 40)
+    private String name;
+
+    @Size(min = 3, max = 15)
+    private String username;
+
+    @Size(max = 40)
+    @Email
+    private String email;
+
+    @Size(min = 6, max = 20)
+    private String password;
+}
