+++ src/main/java/com/list/todo/security/JwtAuthenticationFilter.java
@@ -50,8 +50,10 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
     private String getJwtFromRequest(HttpServletRequest request) {
         String bearerToken = request.getHeader("Authorization");
-        if (StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer ")) {
-            return bearerToken.substring(7, bearerToken.length());
+        String bearer_ = "Bearer ";
+
+        if (StringUtils.hasText(bearerToken) && bearerToken.startsWith(bearer_)) {
+            return bearerToken.substring(bearer_.length());
         }
         return null;
     }
