+++ src/test/java/com/list/todo/controllers/TasksControllerTest.java
@@ -1,0 +1,24 @@
+package com.list.todo.controllers;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class TasksControllerTest {
+
+    @Test
+    public void getAllTasksOnTodoList() {
+    }
+
+    @Test
+    public void addTask() {
+    }
+
+    @Test
+    public void updateTask() {
+    }
+
+    @Test
+    public void deleteTask() {
+    }
+}
